fix reviewSummaryListener
