+++ src/client/pages/Country/Country.tsx
@@ -1,19 +1,16 @@
 import './Country.scss'
-import React, { useEffect } from 'react'
+import React from 'react'
 import { Navigate, Route, Routes } from 'react-router-dom'
 
 import classNames from 'classnames'
 
 import { ClientRoutes } from 'meta/app'
 import { Areas } from 'meta/area'
-import { Sockets } from 'meta/socket'
 import { Authorizer } from 'meta/user'
 
-import { useAppDispatch } from 'client/store'
 import { useCountries, useCountry } from 'client/store/area'
 import { useAssessment, useCycle } from 'client/store/assessment'
 import { useNavigationVisible } from 'client/store/ui/navigation'
-import { ReviewActions } from 'client/store/ui/review'
 import { useUser } from 'client/store/user'
 import { useIsDataExportView } from 'client/hooks'
 import { useCountryRouteParams } from 'client/hooks/useRouteParams'
@@ -22,19 +19,18 @@ import AssessmentDataDownload from 'client/pages/AssessmentDataDownload'
 import AssessmentHome from 'client/pages/AssessmentHome'
 import AssessmentPrint from 'client/pages/AssessmentPrint'
 import AssessmentSection from 'client/pages/AssessmentSection'
-import useGetUsers from 'client/pages/Country/hooks/useGetUsers'
 import DataExport from 'client/pages/DataExport'
 import Geo from 'client/pages/Geo'
 import OriginalDataPoint from 'client/pages/OriginalDataPoint'
 import User from 'client/pages/User'
-import { SocketClient } from 'client/service/socket'
 
+import useGetUsers from './hooks/useGetUsers'
+import { useReviewSummaryListener } from './hooks/useReviewSummaryListener'
 import SectionWrapper from './SectionWrapper'
 
 const Country: React.FC = () => {
   const { assessmentName, cycleName, countryIso } = useCountryRouteParams()
 
-  const dispatch = useAppDispatch()
   const assessment = useAssessment()
   const cycle = useCycle()
   const user = useUser()
@@ -43,36 +39,7 @@ const Country: React.FC = () => {
   const country = useCountry(countryIso)
   const isDataExportView = useIsDataExportView()
   useGetUsers()
-
-  useEffect(() => {
-    return () => {
-      // reset review and assessment section store
-      dispatch(ReviewActions.reset())
-    }
-  }, [countryIso, assessmentName, cycleName, dispatch])
-
-  useEffect(() => {
-    const requestReviewSummaryEvent = Sockets.getRequestReviewSummaryEvent({ countryIso, assessmentName, cycleName })
-
-    const updateReviewSummaryEventHandler = () => {
-      dispatch(ReviewActions.getReviewSummary({ countryIso, assessmentName, cycleName }))
-    }
-
-    if (user) {
-      // fetch review summary
-      dispatch(ReviewActions.getReviewSummary({ countryIso, assessmentName, cycleName }))
-      SocketClient.on(requestReviewSummaryEvent, updateReviewSummaryEventHandler)
-    } else {
-      // reset review summary
-      dispatch(ReviewActions.reset())
-    }
-
-    return () => {
-      if (user) {
-        SocketClient.off(requestReviewSummaryEvent, updateReviewSummaryEventHandler)
-      }
-    }
-  }, [countryIso, assessmentName, cycleName, user, dispatch])
+  useReviewSummaryListener()
 
   if (!countryIso) return null
